S3Object Class

Represents a path or an object in S3, with an optional local copy. S3Objects are not instantiated directly, but they are returned by many methods of the S3Client.

Public fields

exists

Logical indicating if this key corresponds to an object in S3.

url

Character string of the S3 location of the object.

prefix

Character string of the prefix requested that matches this object.

key

Character string corresponding to the key given to the get call that produced this object.

size

Integer size of the object in bytes (or NULL if the key doesn't correspond to an object in S3).

has_info

Logical indicating if this S3Object contains additional metadata.

metadata

List of user-defined metadata (or NULL if no metadata is defined).

content_type

Character string of the content-type of the S3 object (or NULL if not defined).

last_modified

Integer Unix timestamp of the last modified time.

encryption

Character string of the server-side-encryption type (or NULL if not set).

Active bindings

range_info

A list containing information about partial downloads, or NULL.

Methods

Public methods

S3Object$new()

S3Object$print()

S3Object$is_partially_downloaded()

S3Object$clone()

Method new()

Initialize a new S3Object.

Usage

S3Object$new(py_obj)

Arguments

py_obj

A Python S3Object instance.

Method print()

Print method for S3Object.

Usage

S3Object$print(...)

Arguments

...

Additional arguments passed to print (unused).

Method is_partially_downloaded()

Check if the object is partially downloaded.

Usage

S3Object$is_partially_downloaded()

Returns

Logical indicating if the object is partially downloaded.

Method clone()

The objects of this class are cloneable with this method.

Usage

S3Object$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.