Metaflow for R

Metaflow is an open-source framework for data science projects. It's designed to handle the entire lifecycle of data science workflows, from prototyping to production deployment. Metaflow allows you to write and manage data processing pipelines using a simple, intuitive API.

Details

The Metaflow API is primarily implemented in Python, with bindings for R. This package provides access to the complete Metaflow functionality from within R.

Key features of Metaflow include:

Seamless scaling from prototype to production

Built-in versioning of data and code

Easy integration with cloud computing resources

Support for parallel and distributed computing

Documentation

For additional documentation on the Metaflow package, see https://docs.metaflow.org/v/r/

Enterprise Support

For more information about managed Metaflow enterprise offerings, visit Outerbounds

See also

Author

Bryan Galvin (bcgalvin@gmail.com)